requirements.txt
flask==1.0.2
fabric3==1.14.post1
gunicorn==19.8.1
pillow==5.1.0